requirements.txt
-i https://pypi.org/simple
aenum==3.1.15
anyio==4.2.0 ; python_version >= '3.8'
bandwidth-sdk==15.0.1
click==8.1.7 ; python_version >= '3.7'
fastapi==0.109.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.6.1
idna==3.6 ; python_version >= '3.5'
pydantic==1.10.13 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.0
pyyaml==6.0.1
setuptools==69.0.3 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.35.0 ; python_version >= '3.8'
typing-extensions==4.9.0 ; python_version >= '3.8'
urllib3==2.0.7 ; python_version >= '3.7'
uvicorn[standard]==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
